refactor: ソースコードとそのテストをフォルダにまとめる
